Updated files: 0x14-bit_manipulation/1-print_binary.c 0x14-bit_manipulation/_putchar.c 0x14-bit_manipulation/b README.md 0-print_name.c 0x05-pointers_arrays_strings/.1-swap.c.swp 0x05-pointers_arrays_strings/.2-strlen.c.swp function_pointers.h

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -1,6 +1,5 @@
 #include <stddef.h>
 #include "main.h"
-
 /**
  * print_binary - prints the binary representation of a number
  * @n: the number to print
@@ -12,7 +11,7 @@ void print_binary(unsigned long int n)
 	unsigned long int mask = 1;
 	int i;
 
-	for (i = 0; i < (sizeof(unsigned long int) * 8); i++)
+	for (i = 0; i < (int)(sizeof(unsigned long int) * 8); i++)
 	{
 		if ((n >> i) == 0)
 			break;
